+++ test-utils/multiple-database-version-helper.js
@@ -4,6 +4,7 @@ const {
   DATABASE_URL_MSSQL,
   DATABASE_URL_MYSQL,
   DATABASE_URL_POSTGRESQL,
+  DATABASE_URL_POSTGRESQL_OLD,
 } = require('./database-urls');
 
 const mongoDatabases = [{
@@ -16,14 +17,22 @@ const mongoDatabases = [{
 
 const sqlDatabases = [{
   dialect: 'mysql',
+  version: '5.7',
   connectionUrl: DATABASE_URL_MYSQL,
   schema: 'public',
 }, {
   dialect: 'postgres',
+  version: '9.4',
+  connectionUrl: DATABASE_URL_POSTGRESQL_OLD,
+  schema: 'public',
+}, {
+  dialect: 'postgres',
+  version: '12.1',
   connectionUrl: DATABASE_URL_POSTGRESQL,
   schema: 'public',
 }, {
   dialect: 'mssql',
+  version: '2017-CU8-ubuntu',
   connectionUrl: DATABASE_URL_MSSQL,
   schema: 'dbo',
 }];
@@ -38,7 +47,7 @@ module.exports = {
   describeSequelizeDatabases(tests) {
     sqlDatabases.forEach((sqlDatabase) => {
       // eslint-disable-next-line jest/valid-describe
-      describe(`using ${sqlDatabase.dialect} Database`, tests(sqlDatabase));
+      describe(`using ${sqlDatabase.dialect} Database v${sqlDatabase.version}`, tests(sqlDatabase));
     });
   },
 };